refactor : 예약 관련 연관관계 재설정(다대다 -> 일대일) 및 예약 관련 JPA 구현체 생성
