--- src/main/java/roomescape/domain/reservation/domain/ReservationMember.java
@@ -1,20 +1,0 @@
-package roomescape.domain.reservation.domain;
-
-import jakarta.persistence.*;
-import roomescape.domain.member.domain.Member;
-
-@Entity
-public class ReservationMember {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "reservation_id")
-    private Reservation reservation;
-
-    @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-}
