+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -5,7 +5,9 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.member.domain.Member;
 import roomescape.domain.member.service.MemberService;
 import roomescape.domain.reservation.domain.Reservation;
-import roomescape.domain.reservation.domain.repository.ReservationRepository;
+import roomescape.domain.reservation.domain.repository.ReservationJpaRepository;
+import roomescape.domain.reservation.error.exception.ReservationErrorCode;
+import roomescape.domain.reservation.error.exception.ReservationException;
 import roomescape.domain.reservation.service.dto.AdminReservationRequest;
 import roomescape.domain.reservation.service.dto.ReservationRequest;
 import roomescape.domain.theme.domain.Theme;
@@ -15,15 +17,19 @@ import roomescape.domain.time.service.TimeService;
 
 import java.util.List;
 
+import static roomescape.utils.DateTimeCheckUtil.isBeforeCheck;
+import static roomescape.utils.FormatCheckUtil.reservationDateFormatCheck;
+import static roomescape.utils.FormatCheckUtil.reservationNameFormatCheck;
+
 @Service
 public class ReservationService {
 
     private final TimeService timeService;
     private final ThemeService themeService;
     private final MemberService memberService;
-    private final ReservationRepository reservationRepository;
+    private final ReservationJpaRepository reservationRepository;
 
-    public ReservationService(TimeService timeService, ThemeService themeService, MemberService memberService, ReservationRepository reservationRepository) {
+    public ReservationService(TimeService timeService, ThemeService themeService, MemberService memberService, ReservationJpaRepository reservationRepository) {
         this.timeService = timeService;
         this.themeService = themeService;
         this.memberService = memberService;
@@ -34,9 +40,9 @@ public class ReservationService {
     public Reservation save(ReservationRequest reservationRequest, Member loginMember) {
         Time time = timeService.findById(reservationRequest.getTimeId());
         Theme theme = themeService.findById(reservationRequest.getThemeId());
-        Reservation reservation = new Reservation(null, reservationRequest.getName(), reservationRequest.getDate(), time, theme, loginMember);
-        Long id = reservationRepository.save(reservation);
-        return findById(id);
+        validationCheck(reservationRequest.getName(), reservationRequest.getDate(), time);
+        Reservation reservation = new Reservation(null, reservationRequest.getName(), reservationRequest.getDate(), theme, time, loginMember);
+        return reservationRepository.save(reservation);
     }
 
     @Transactional
@@ -44,14 +50,13 @@ public class ReservationService {
         Time time = timeService.findById(adminReservationRequest.getTimeId());
         Theme theme = themeService.findById(adminReservationRequest.getThemeId());
         Member member = memberService.findById(adminReservationRequest.getMemberId());
-        Reservation reservation = new Reservation(null, member.getName(), adminReservationRequest.getDate(), time, theme, member);
-        Long id = reservationRepository.save(reservation);
-        return findById(id);
+        Reservation reservation = new Reservation(null, member.getName(), adminReservationRequest.getDate(), theme, time, member);
+        return reservationRepository.save(reservation);
     }
 
     @Transactional(readOnly = true)
     public Reservation findById(long id) {
-        return reservationRepository.findById(id);
+        return reservationRepository.findById(id).orElseThrow(() -> new ReservationException(ReservationErrorCode.INVALID_RESERVATION_DETAILS_ERROR));
     }
 
     @Transactional(readOnly = true)
@@ -61,6 +66,13 @@ public class ReservationService {
 
     @Transactional
     public void delete(Long id) {
-        reservationRepository.delete(id);
+        Reservation reservation = reservationRepository.findById(id).orElseThrow(() -> new ReservationException(ReservationErrorCode.INVALID_RESERVATION_DETAILS_ERROR));
+        reservationRepository.delete(reservation);
+    }
+
+    private void validationCheck(String name, String date, Time time) {
+        reservationNameFormatCheck(name);
+        reservationDateFormatCheck(date);
+        isBeforeCheck(date, time.getStartAt());
     }
 }
